refactor(GameClass): add namespace GameObj.Bullets
- split BulletFactory
- split Laser, Plasma, Shell, Missile, Arc
- remove the braces of namespace
- refactor AP
